Fixed cd bug, run if invoked directly, use $PWD for add
There was some bug when changing to the project's directory.
I don't remember what it was, but it's fixed.

If the project.sh is ran directly instead of sourced, it will execute
the project function. Obviously "project cd" won't work.

If no path is given to "project add" it will use the current
directory. 9 times out of 10 I was $PWD for the path so this
is one less thing to type.